refactor: split recovery key display
